backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langgraph==0.0.20
langchain==0.1.0
langchain-openai==0.0.5
langchain-core>=0.1.8,<0.2.0
openai>=1.12.0
python-dotenv==1.0.0
GitPython==3.1.41
httpx==0.24.1
sqlalchemy>=2.0.46
aiosqlite==0.19.0
